+++ tests/modules/test_SeleniumExecutor.py
@@ -1,4 +1,5 @@
 import os
+import re
 import shutil
 import time
 
@@ -19,6 +20,12 @@ class SeleniumTestCase(BZTestCase):
         self.engine_obj.config.get("modules").get("selenium").merge({"virtual-display": {"width": 1024, "height": 768}})
         self.selenium_config = self.engine_obj.config["modules"]["selenium"]
         self.engine_obj.create_artifacts_dir(self.paths)
+        self.obj = SeleniumExecutor()
+        self.obj.engine = self.engine_obj
+        self.obj.settings = self.selenium_config
+
+    def tearDown(self):
+        self.obj.free_virtual_display()
 
 
 class TestSeleniumJUnitRunner(SeleniumTestCase):
@@ -45,19 +52,18 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
 
         self.assertFalse(os.path.exists(dummy_installation_path))
 
-        obj = self.get_selenium_executor()
-        obj.settings.merge({"selenium-tools": {
+        self.obj.settings.merge({"selenium-tools": {
             "junit": {"selenium-server": os.path.join(dummy_installation_path, "selenium-server.jar")}
         }})
-        obj.settings.merge({"selenium-tools": {
+        self.obj.settings.merge({"selenium-tools": {
             "junit": {"hamcrest-core": os.path.join(dummy_installation_path, "tools", "junit", "hamcrest-core.jar")}
         }})
-        obj.settings.merge({"selenium-tools": {
+        self.obj.settings.merge({"selenium-tools": {
             "junit": {"path": os.path.join(dummy_installation_path, "tools", "junit", "junit.jar")}
         }})
 
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/"}})
-        obj.prepare()
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/"}})
+        self.obj.prepare()
         self.assertTrue(os.path.exists(os.path.join(dummy_installation_path, "selenium-server.jar")))
         self.assertTrue(os.path.exists(os.path.join(dummy_installation_path, "tools", "junit", "junit.jar")))
         self.assertTrue(os.path.exists(os.path.join(dummy_installation_path, "tools", "junit", "hamcrest-core.jar")))
@@ -66,33 +72,25 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         SeleniumExecutor.HAMCREST_DOWNLOAD_LINK = hamcrest_link
         SeleniumExecutor.JUNIT_MIRRORS_SOURCE = junit_mirrors
 
-    def get_selenium_executor(self):
-        obj = SeleniumExecutor()
-        obj.engine = self.engine_obj
-        obj.settings = self.selenium_config
-        return obj
-
     def test_prepare_java_single(self):
         """
         Check if script exists in working dir
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java/TestBlazemeterFail.java"}})
-        obj.prepare()
-        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "TestBlazemeterFail.java")))
-        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "TestBlazemeterFail.class")))
-        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "compiled.jar")))
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java/TestBlazemeterFail.java"}})
+        self.obj.prepare()
+        self.assertTrue(os.path.exists(os.path.join(self.obj.runner.working_dir, "TestBlazemeterFail.java")))
+        self.assertTrue(os.path.exists(os.path.join(self.obj.runner.working_dir, "TestBlazemeterFail.class")))
+        self.assertTrue(os.path.exists(os.path.join(self.obj.runner.working_dir, "compiled.jar")))
 
     def test_prepare_java_folder(self):
         """
         Check if scripts exist in working dir
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java/"}})
-        obj.prepare()
-        prepared_files = os.listdir(obj.runner.working_dir)
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java/"}})
+        self.obj.prepare()
+        prepared_files = os.listdir(self.obj.runner.working_dir)
         java_files = [fname for fname in prepared_files if fname.endswith(".java")]
         class_files = [fname for fname in prepared_files if fname.endswith(".class")]
         jars = [fname for fname in prepared_files if fname.endswith(".jar")]
@@ -105,53 +103,49 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         Check if scripts exist in working dir
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java_package/"}})
-        obj.prepare()
-        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "compiled.jar")))
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java_package/"}})
+        self.obj.prepare()
+        self.assertTrue(os.path.exists(os.path.join(self.obj.runner.working_dir, "compiled.jar")))
 
     def test_selenium_startup_shutdown_java_package(self):
         """
         Run tests from package
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/java_package/'},
                 'executor': 'selenium'
             },
             'reporting': [{'module': 'junit-xml'}]
         })
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
-        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "compiled.jar")))
+        self.obj.shutdown()
+        self.assertTrue(os.path.exists(os.path.join(self.obj.runner.working_dir, "compiled.jar")))
 
     def test_prepare_jar_single(self):
         """
         Check if jar exists in working dir
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/dummy.jar"}})
-        obj.prepare()
-        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "dummy.jar")))
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/dummy.jar"}})
+        self.obj.prepare()
+        self.assertTrue(os.path.exists(os.path.join(self.obj.runner.working_dir, "dummy.jar")))
 
     def test_prepare_jar_folder(self):
         """
         Check if jars exist in working dir
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/"}})
-        obj.prepare()
-        java_scripts = os.listdir(obj.runner.working_dir)
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/"}})
+        self.obj.prepare()
+        java_scripts = os.listdir(self.obj.runner.working_dir)
         self.assertEqual(len(java_scripts), 2)
 
     def test_selenium_startup_shutdown_jar_single(self):
@@ -159,104 +153,100 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
         runt tests from single jar
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/jar/'},
                 'executor': 'selenium'
             },
             'reporting': [{'module': 'junit-xml'}]
         })
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/dummy.jar"}})
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/jar/dummy.jar"}})
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
+        self.obj.shutdown()
 
-        prepared_files = os.listdir(obj.runner.working_dir)
+        prepared_files = os.listdir(self.obj.runner.working_dir)
         java_files = [fname for fname in prepared_files if fname.endswith(".java")]
         class_files = [fname for fname in prepared_files if fname.endswith(".class")]
         jars = [fname for fname in prepared_files if fname.endswith(".jar")]
         self.assertEqual(len(java_files), 0)
         self.assertEqual(len(class_files), 0)
         self.assertEqual(len(jars), 1)
-        self.assertTrue(os.path.exists(obj.runner.settings.get("report-file")))
+        self.assertTrue(os.path.exists(self.obj.runner.settings.get("report-file")))
 
     def test_selenium_startup_shutdown_jar_folder(self):
         """
         run tests from jars
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/jar/'},
                 'executor': 'selenium'
             },
             'reporting': [{'module': 'junit-xml'}]
         })
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
+        self.obj.shutdown()
 
-        prepared_files = os.listdir(obj.runner.working_dir)
+        prepared_files = os.listdir(self.obj.runner.working_dir)
         java_files = [fname for fname in prepared_files if fname.endswith(".java")]
         class_files = [fname for fname in prepared_files if fname.endswith(".class")]
         jars = [fname for fname in prepared_files if fname.endswith(".jar")]
         self.assertEqual(len(java_files), 0)
         self.assertEqual(len(class_files), 0)
         self.assertEqual(len(jars), 2)
-        self.assertTrue(os.path.exists(obj.runner.settings.get("report-file")))
+        self.assertTrue(os.path.exists(self.obj.runner.settings.get("report-file")))
 
     def test_selenium_startup_shutdown_java_single(self):
         """
         run tests from single .java file
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/java/'},
                 'executor': 'selenium'
             },
             'reporting': [{'module': 'junit-xml'}]
         })
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java/TestBlazemeterFail.java"}})
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/java/TestBlazemeterFail.java"}})
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
+        self.obj.shutdown()
 
-        prepared_files = os.listdir(obj.runner.working_dir)
+        prepared_files = os.listdir(self.obj.runner.working_dir)
         java_files = [fname for fname in prepared_files if fname.endswith(".java")]
         class_files = [fname for fname in prepared_files if fname.endswith(".class")]
         jars = [fname for fname in prepared_files if fname.endswith(".jar")]
         self.assertEqual(1, len(java_files))
         self.assertEqual(1, len(class_files))
         self.assertEqual(1, len(jars))
-        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "compiled.jar")))
-        self.assertTrue(os.path.exists(obj.runner.settings.get("report-file")))
+        self.assertTrue(os.path.exists(os.path.join(self.obj.runner.working_dir, "compiled.jar")))
+        self.assertTrue(os.path.exists(self.obj.runner.settings.get("report-file")))
 
     def test_selenium_startup_shutdown_java_folder(self):
         """
         run tests from .java files
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/java/'},
                 'executor': 'selenium'
@@ -264,93 +254,88 @@ class TestSeleniumJUnitRunner(SeleniumTestCase):
             'reporting': [{'module': 'junit-xml'}]
         })
 
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
+        self.obj.shutdown()
 
-        prepared_files = os.listdir(obj.runner.working_dir)
+        prepared_files = os.listdir(self.obj.runner.working_dir)
         java_files = [fname for fname in prepared_files if fname.endswith(".java")]
         class_files = [fname for fname in prepared_files if fname.endswith(".class")]
         jars = [fname for fname in prepared_files if fname.endswith(".jar")]
         self.assertEqual(2, len(java_files))
         self.assertEqual(2, len(class_files))
         self.assertEqual(1, len(jars))
-        self.assertTrue(os.path.exists(os.path.join(obj.runner.working_dir, "compiled.jar")))
-        self.assertTrue(os.path.exists(obj.runner.settings.get("report-file")))
+        self.assertTrue(os.path.exists(os.path.join(self.obj.runner.working_dir, "compiled.jar")))
+        self.assertTrue(os.path.exists(self.obj.runner.settings.get("report-file")))
 
     def test_not_junit(self):
         """
         Check that JUnit runner fails if no tests were found
         :return:
         """
-        obj = self.get_selenium_executor()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             Provisioning.PROV: "local",
             ScenarioExecutor.EXEC: {
                 "executor": "selenium",
                 "scenario": {"script": __dir__() + "/../selenium/invalid/NotJUnittest.java"}
             }
         })
-        obj.execution = obj.engine.config['execution']
-        obj.prepare()
-        obj.startup()
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.prepare()
+        self.obj.startup()
         try:
-            while not obj.check():
+            while not self.obj.check():
                 time.sleep(1)
             self.fail()
         except BaseException as exc:
             self.assertIn("Nothing to test", exc.args[0])
-        obj.shutdown()
+        self.obj.shutdown()
 
     def test_resource_files_collection_remote_java(self):
-        obj = self.get_selenium_executor()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/java/'},
                 'executor': 'selenium'
             },
             'reporting': [{'module': 'junit-xml'}]
         })
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
 
-        self.assertEqual(len(obj.resource_files()), 1)
+        self.assertEqual(len(self.obj.resource_files()), 1)
 
     def test_resource_files_collection_remote_jar(self):
-        obj = self.get_selenium_executor()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/jar/'},
                 'executor': 'selenium'
             },
             'reporting': [{'module': 'junit-xml'}]
         })
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
 
-        self.assertEqual(len(obj.resource_files()), 1)
+        self.assertEqual(len(self.obj.resource_files()), 1)
 
 
-class TestSeleniumNoseRunner(BZTestCase):
+class TestSeleniumNoseRunner(SeleniumTestCase):
     def test_selenium_prepare_python_single(self):
         """
         Check if script exists in working dir
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.execution.merge({"scenario": {
+        self.obj.execution.merge({"scenario": {
             "script": __dir__() + "/../selenium/python/test_blazemeter_fail.py"
         }})
-        obj.prepare()
-        python_scripts = os.listdir(obj.runner.working_dir)
+        self.obj.prepare()
+        python_scripts = os.listdir(self.obj.runner.working_dir)
         self.assertEqual(len(python_scripts), 1)
 
     def test_selenium_prepare_python_folder(self):
@@ -358,11 +343,9 @@ class TestSeleniumNoseRunner(BZTestCase):
         Check if scripts exist in working dir
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/python/"}})
-        obj.prepare()
-        python_scripts = os.listdir(obj.runner.working_dir)
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/python/"}})
+        self.obj.prepare()
+        python_scripts = os.listdir(self.obj.runner.working_dir)
         self.assertEqual(len(python_scripts), 2)
 
     def test_selenium_startup_shutdown_python_single(self):
@@ -371,91 +354,83 @@ class TestSeleniumNoseRunner(BZTestCase):
         :return:
         """
 
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/python/'},
                 'executor': 'selenium'
             },
             'reporting': [{'module': 'junit-xml'}]
         })
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
 
-        obj.execution.merge({"scenario": {
+        self.obj.execution.merge({"scenario": {
             "script": __dir__() + "/../selenium/python/test_blazemeter_fail.py"
         }})
 
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
-        prepared_files = os.listdir(obj.runner.working_dir)
+        self.obj.shutdown()
+        prepared_files = os.listdir(self.obj.runner.working_dir)
         python_files = [fname for fname in prepared_files if fname.endswith(".py")]
         self.assertEqual(1, len(python_files))
-        self.assertTrue(os.path.exists(obj.runner.settings.get("report-file")))
+        self.assertTrue(os.path.exists(self.obj.runner.settings.get("report-file")))
 
     def test_selenium_startup_shutdown_python_folder(self):
         """
         run tests from .py files
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             'execution': {
                 'scenario': {'script': __dir__() + '/../selenium/python/'},
                 'executor': 'selenium'
             },
             'reporting': [{'module': 'junit-xml'}]
         })
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
-        prepared_files = os.listdir(obj.runner.working_dir)
+        self.obj.shutdown()
+        prepared_files = os.listdir(self.obj.runner.working_dir)
         python_files = [fname for fname in prepared_files if fname.endswith(".py")]
         self.assertEqual(2, len(python_files))
-        self.assertTrue(os.path.exists(obj.runner.settings.get("report-file")))
+        self.assertTrue(os.path.exists(self.obj.runner.settings.get("report-file")))
 
     def runner_fail_no_test_found(self):
         """
         Check that Python Nose runner fails if no tests were found
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             ScenarioExecutor.EXEC: {
                 "executor": "selenium",
                 "scenario": {"script": __dir__() + "/../selenium/invalid/dummy.py"}
             }
         })
-        obj.execution = obj.engine.config['execution']
-        obj.prepare()
-        obj.startup()
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.prepare()
+        self.obj.startup()
         try:
-            while not obj.check():
+            while not self.obj.check():
                 time.sleep(1)
             self.fail()
         except RuntimeError as exc:
             self.assertIn("Nothing to test.", exc.args[0])
-        obj.shutdown()
+        self.obj.shutdown()
 
     def test_resource_files_collection_remote_nose(self):
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/python/"}})
-        obj.settings.merge(obj.engine.config.get("modules").get("selenium"))
+        self.obj.execution.merge({"scenario": {"script": __dir__() + "/../selenium/python/"}})
+        self.obj.settings.merge(self.obj.engine.config.get("modules").get("selenium"))
 
-        self.assertEqual(len(obj.resource_files()), 1)
+        self.assertEqual(len(self.obj.resource_files()), 1)
 
 
 class TestSeleniumStuff(SeleniumTestCase):
@@ -464,130 +439,109 @@ class TestSeleniumStuff(SeleniumTestCase):
         Raise runtime error when no scenario provided
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.engine.config.merge({ScenarioExecutor.EXEC: {"executor": "selenium"}})
-        obj.execution = obj.engine.config['execution']
-        self.assertRaises(ValueError, obj.prepare)
+        self.obj.engine.config.merge({ScenarioExecutor.EXEC: {"executor": "selenium"}})
+        self.obj.execution = self.obj.engine.config['execution']
+        self.assertRaises(ValueError, self.obj.prepare)
 
     def test_javac_fail(self):
         """
         Test RuntimeError when compilation fails
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = self.engine_obj
-        obj.settings = self.selenium_config
-        obj.engine.config.merge({
+        self.obj.engine.config.merge({
             ScenarioExecutor.EXEC: {
                 "executor": "selenium",
                 "scenario": {"script": __dir__() + "/../selenium/invalid/invalid.java"}
             }
         })
-        obj.execution = obj.engine.config['execution']
-        self.assertRaises(RuntimeError, obj.prepare)
+        self.obj.execution = self.obj.engine.config['execution']
+        self.assertRaises(RuntimeError, self.obj.prepare)
 
     def test_no_supported_files_to_test(self):
         """
         Test RuntimeError raised when no files of known types were found.
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.engine.config.merge({ScenarioExecutor.EXEC: {
+        self.obj.engine.config.merge({ScenarioExecutor.EXEC: {
             "executor": "selenium",
             "scenario": {"script": __dir__() + "/../selenium/invalid/not_found"}
         }})
-        obj.execution = obj.engine.config['execution']
-        self.assertRaises(ValueError, obj.prepare)
+        self.obj.execution = self.obj.engine.config['execution']
+        self.assertRaises(ValueError, self.obj.prepare)
 
     def test_samples_count_annotations(self):
         """
         Test exact number of tests when java annotations used
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = self.engine_obj
-        obj.settings = self.selenium_config
-        obj.engine.config.merge({ScenarioExecutor.EXEC: {
+        self.obj.settings = self.selenium_config
+        self.obj.engine.config.merge({ScenarioExecutor.EXEC: {
             "executor": "selenium",
             "scenario": {"script": __dir__() + "/../selenium/invalid/SeleniumTest.java"}
         }})
-        obj.execution = obj.engine.config['execution']
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
+        self.obj.shutdown()
 
     def test_samples_count_testcase(self):
         """
         Test exact number of tests when test class extends JUnit TestCase
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = self.engine_obj
-        obj.settings = self.selenium_config
-        obj.engine.config.merge({ScenarioExecutor.EXEC: {
+        self.obj.engine.config.merge({ScenarioExecutor.EXEC: {
             "executor": "selenium",
             "scenario": {"script": __dir__() + "/../selenium/invalid/SimpleTest.java"}
         }})
-        obj.execution = obj.engine.config['execution']
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
+        self.obj.shutdown()
 
     def test_no_test_in_name(self):
         """
         Test exact number of tests when annotations used and no "test" in class name
         :return:
         """
-        obj = SeleniumExecutor()
-        obj.engine = self.engine_obj
-        obj.settings = self.selenium_config
-        obj.engine.config.merge({ScenarioExecutor.EXEC: {
+        self.obj.engine.config.merge({ScenarioExecutor.EXEC: {
             "executor": "selenium",
             "scenario": {"script": __dir__() + "/../selenium/invalid/selenium1.java"}
         }})
-        obj.execution = obj.engine.config['execution']
-        obj.prepare()
-        obj.startup()
-        while not obj.check():
+        self.obj.execution = self.obj.engine.config['execution']
+        self.obj.prepare()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
+        self.obj.shutdown()
 
     def test_requests(self):
-        obj = SeleniumExecutor()
-        obj.engine = self.engine_obj
-        obj.settings = self.selenium_config
-        obj.engine.config.merge(yaml.load(open(__dir__() + "/../yaml/selenium_executor_requests.yml").read()))
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
+        self.obj.engine.config.merge(yaml.load(open(__dir__() + "/../yaml/selenium_executor_requests.yml").read()))
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
 
-        obj.prepare()
-        obj.get_widget()
-        obj.startup()
-        while not obj.check():
+        self.obj.prepare()
+        self.obj.get_widget()
+        self.obj.startup()
+        while not self.obj.check():
             time.sleep(1)
-        obj.shutdown()
-        with open(os.path.join(obj.engine.artifacts_dir, "junit.err")) as fds:
+        self.obj.shutdown()
+        with open(os.path.join(self.obj.engine.artifacts_dir, "junit.err")) as fds:
             contents = fds.read()
             self.assertEqual(1, contents.count("ok"), "file: '%s', size: %s, content: '%s'" % (fds, fds.__sizeof__(),
                                                                                                contents))
             self.assertEqual(1, contents.count("OK"))
 
     def test_fail_on_zero_results(self):
-        obj = SeleniumExecutor()
-        obj.engine = self.engine_obj
-        obj.settings = self.selenium_config
-        obj.engine.config.merge(yaml.load(open(__dir__() + "/../yaml/selenium_executor_requests.yml").read()))
-        obj.engine.config.merge({"provisioning": "local"})
-        obj.execution = obj.engine.config['execution']
+        self.obj.engine.config.merge(yaml.load(open(__dir__() + "/../yaml/selenium_executor_requests.yml").read()))
+        self.obj.engine.config.merge({"provisioning": "local"})
+        self.obj.execution = self.obj.engine.config['execution']
 
-        obj.prepare()
-        self.assertRaises(RuntimeWarning, obj.post_process)
+        self.obj.prepare()
+        self.assertRaises(RuntimeWarning, self.obj.post_process)
 
     def test_junit_mirrors(self):
         dummy_installation_path = __dir__() + "/../../../build/tmp/selenium-taurus"
@@ -598,59 +552,66 @@ class TestSeleniumStuff(SeleniumTestCase):
         objjm.install()
 
     def test_remote_prov_requests(self):
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.execution.merge({
+        self.obj.execution.merge({
             "scenario": {
                 "requests": [
                     "http://blazedemo.com"
                 ]
             }
         })
-        obj.resource_files()
+        self.obj.resource_files()
 
     def test_dont_copy_local_script_to_artifacts(self):
         "ensures that .java file is not copied into artifacts-dir"
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
         filename = "BlazeDemo.java"
         script_path = __dir__() + "/../data/" + filename
-        obj.execution.merge({
+        self.obj.execution.merge({
             "scenario": {
                 "script": script_path,
             }
         })
-        obj.prepare()
-        files = obj.resource_files()
+        self.obj.prepare()
+        files = self.obj.resource_files()
         self.assertIn(script_path, files)
-        artifacts_script = os.path.join(obj.engine.artifacts_dir, filename)
+        artifacts_script = os.path.join(self.obj.engine.artifacts_dir, filename)
         self.assertFalse(os.path.exists(artifacts_script))
 
     def test_take_script_from_artifacts(self):
         "ensures that executor looks for script in artifacts-dir (for cloud/remote cases)"
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.engine.file_search_paths = [obj.engine.artifacts_dir]
+        self.obj.engine.file_search_paths = [self.obj.engine.artifacts_dir]
 
         script_name = "BlazeDemo.java"
         test_script = __dir__() + "/../data/" + script_name
-        artifacts_script = os.path.join(obj.engine.artifacts_dir, script_name)
+        artifacts_script = os.path.join(self.obj.engine.artifacts_dir, script_name)
         shutil.copy2(test_script, artifacts_script)
 
-        obj.execution.merge({
+        self.obj.execution.merge({
             "scenario": {
                 "script": script_name,
             }
         })
-        obj.prepare()
+        self.obj.prepare()
 
     def test_do_not_modify_scenario_script(self):
-        obj = SeleniumExecutor()
-        obj.engine = EngineEmul()
-        obj.execution.merge({
+        self.obj.execution.merge({
             "scenario": {
                 "requests": ["address"],
             }
         })
-        obj.prepare()
-        self.assertNotIn("script", obj.get_scenario())
+        self.obj.prepare()
+        self.assertNotIn("script", self.obj.get_scenario())
+
+    def test_default_address_gen(self):
+        self.obj.execution.merge({
+            "scenario": {
+                "default-address": "http://blazedemo.com",
+                "requests": ["/", "http://absolute.address.com/somepage","/reserve.php"],
+            }
+        })
+        self.obj.prepare()
+        with open(os.path.join(self.obj.runner_working_dir, os.path.basename(self.obj.script))) as fds:
+            script = fds.read()
+        urls = re.findall(r"get\('(.+)'\)", script)
+        self.assertEqual("http://blazedemo.com/", urls[0])
+        self.assertEqual("http://absolute.address.com/somepage", urls[1])
+        self.assertEqual("http://blazedemo.com/reserve.php", urls[2])