+++ site/dat/docs/Changelog.md
@@ -5,6 +5,7 @@
  - fix 'non-existent scenario' error case
  - fix config cleanup in cloud provisioning
  - fix Selenium crash when used in multi-execution with a shared scenario
+ - add `default-address` scenario option to Selenium
 
 ## 1.6.3 <sup>17 jun 2016</sup>
  - fix percentile value handling in passfail criteria